Deployment/scripts/index_scripts/requirements.txt
azure-storage-file-datalake
# langchain
openai
pypdf
# pyodbc
tiktoken
msal[broker]==1.24.0b1
azure-identity
azure-ai-textanalytics
azure-search-documents==11.6.0b3
azure-keyvault-secrets
pandas
pymssql
datetime
# graphrag==0.3.6